refactor(mobile): extract BackgroundBackupQueuePort/Result to break import cycle
